+++ packages/monika-notification/channel/teams.ts
@@ -23,8 +23,8 @@
  **********************************************************************************/
 
 import Joi from 'joi'
-import { sendHttpRequest } from '../../../utils/http'
-import type { NotificationMessage } from './'
+import { sendHttpRequest } from '../utils/http'
+import type { NotificationMessage } from '.'
 
 type NotificationData = {
   url: string
