+++ packages/monika-notification/channel/telegram.ts
@@ -25,7 +25,7 @@
 /* eslint-disable camelcase */
 import Joi from 'joi'
 import type { NotificationMessage } from './'
-import { sendHttpRequest } from '../../../utils/http'
+import { sendHttpRequest } from '../utils/http'
 
 type NotificationData = {
   group_id: string
