+++ packages/monika-notification/channel/workplace.ts
@@ -1,0 +1,60 @@
+/**********************************************************************************
+ * MIT License                                                                    *
+ *                                                                                *
+ * Copyright (c) 2021 Hyperjump Technology                                        *
+ *                                                                                *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy   *
+ * of this software and associated documentation files (the "Software"), to deal  *
+ * in the Software without restriction, including without limitation the rights   *
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell      *
+ * copies of the Software, and to permit persons to whom the Software is          *
+ * furnished to do so, subject to the following conditions:                       *
+ *                                                                                *
+ * The above copyright notice and this permission notice shall be included in all *
+ * copies or substantial portions of the Software.                                *
+ *                                                                                *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR     *
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,       *
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE    *
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER         *
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,  *
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE  *
+ * SOFTWARE.                                                                      *
+ **********************************************************************************/
+
+/* eslint-disable camelcase */
+import Joi from 'joi'
+import type { NotificationMessage } from '.'
+import { sendHttpRequest } from '../utils/http'
+
+type NotificationData = {
+  thread_id: string
+  access_token: string
+}
+
+export const validator = Joi.object().keys({
+  thread_id: Joi.string().required().label('Workplace Thread ID'),
+  access_token: Joi.string().required().label('Workplace Access Token'),
+})
+
+export const send = async (
+  { access_token, thread_id }: NotificationData,
+  { body }: NotificationMessage
+): Promise<void> => {
+  await sendHttpRequest({
+    baseURL: 'https://graph.workplace.com',
+    headers: {
+      Authorization: `Bearer ${access_token}`,
+    },
+    method: 'POST',
+    url: '/me/messages',
+    data: {
+      recipient: {
+        thread_key: thread_id,
+      },
+      message: {
+        text: body,
+      },
+    },
+  })
+}
